Move initialization code out of FileSystemServer
