+++ src/LibHac/Fs/SaveData.cs
@@ -1,0 +1,7 @@
+﻿namespace LibHac.Fs
+{
+    public static class SaveData
+    {
+        public const ulong SaveIndexerId = 0x8000000000000000;
+    }
+}
