+++ tests/LibHac.Tests/Fs/FileSystemClientTests/FileSystemServerFactory.cs
@@ -20,7 +20,7 @@ namespace LibHac.Tests.Fs.FileSystemClientTests
             config.DeviceOperator = defaultObjects.DeviceOperator;
             config.ExternalKeySet = new ExternalKeySet();
 
-            var horizon = new Horizon(new StopWatchTimeSpanGenerator(), config);
+            Horizon horizon = LibHac.HorizonFactory.CreateWithFsConfig(new StopWatchTimeSpanGenerator(), config);
 
             HorizonClient horizonClient = horizon.CreatePrivilegedHorizonClient();
